Implement a simple resampler working only for input_freq / output_freq in range [0.5 1.0].
It use a linear interpolation plus a lowpass filter. This is used by the
macplus emulator with it's 22255 hz frequency which is actually mapped to
32000 hz by the sdl1 emscripten port. It use floating point index in the
input buffer because 22255 = 4451 * 5 and 4451 is prime which make
difficult to implement it through a multi pass insertion decimation.

This applied to Hampa Hug work well but in a browser there is still some
artifact at buffer boundary. It's unclear if it comes from the browser
unable to call enough fast the buffer filler or from the rewrite of the
main emulator loop.